requirements.txt
flask
line-bot-sdk
requests
openai
python-dotenv
